+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Added
 - Unit tests for FunFair.BuildCheck.Helpers to achieve 100% code coverage
 - Unit tests for FunFair.BuildCheck.SolutionChecks to achieve 100% code coverage
+- Unit tests for FunFair.BuildCheck.Runner to achieve 100% code coverage
 ### Fixed
 ### Changed
 ### Deprecated